requirements.txt
requests
python_dotenv
progress
